test: 新增手写测试用例
