+++ src/skeleton/__tests__/skeleton.test.jsx
@@ -1,0 +1,43 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+/**
+ * 该文件由脚本自动生成，如需修改请联系 PMC
+ * This file generated by scripts of tdesign-api. `npm run api:docs Skeleton React(PC) vitest,finalProject`
+ * If you need to modify this file, contact PMC first please.
+ */
+import React from 'react';
+import { fireEvent, vi, render, mockDelay, screen } from '@test/utils';
+import { Skeleton } from '..';
+import { getSkeletonDefaultMount, getSkeletonInfo } from './mount';
+
+describe('Skeleton Component', () => {
+  it('props.delay: show loading delay 10ms', async () => {
+    const props = { delay: 10 };
+    const { mountedContent, getContent } = getSkeletonInfo(Skeleton, props);
+    expect(mountedContent.container.querySelector('.t-skeleton__row')).toBeFalsy();
+    // await mockDelay(10);
+    // mountedContent.rerender(getContent(Skeleton, {...props, loading: false}));
+    // expect(mountedContent.container.querySelector('.t-skeleton__row')).toBeFalsy();
+    // await mockDelay(20);
+    // expect(mountedContent.container.querySelector('.t-skeleton__row')).toBeFalsy();
+
+    // mountedContent.rerender(getContent(Skeleton, { ...props, loading: true }));
+    // // <34 都不会有骨架屏
+    // await mockDelay(35);
+    // mountedContent.rerender(getContent(Skeleton, { ...props, loading: false }));
+    // expect(mountedContent.container.querySelector('.t-skeleton__row')).toBeTruthy();
+
+    setTimeout(() => { 
+      mountedContent.rerender(getContent(Skeleton, {...props, loading: false}));
+      expect(mountedContent.container.querySelector('.t-skeleton__row')).toBeFalsy();
+    }, 10);
+    setTimeout(() => { 
+      expect(mountedContent.container.querySelector('.t-skeleton__row')).toBeFalsy();
+      mountedContent.rerender(getContent(Skeleton, { ...props, loading: true }));
+    }, 20);
+    
+    // <34 都不会有骨架屏
+    await mockDelay(35);
+    mountedContent.rerender(getContent(Skeleton, { ...props, loading: false }));
+    expect(mountedContent.container.querySelector('.t-skeleton__row')).toBeTruthy();
+  });
+});
